Fix cluster-autoscaler Application rendering
